Fix reading from some ZIP files
