+++ package.json
@@ -40,7 +40,7 @@
     "del-cli": "^6.0.0",
     "file-type": "^19.6.0",
     "mocha": "^10.8.2",
-    "strtok3": "^10.0.0",
+    "strtok3": "^10.0.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.7.2"
   },
